fix: suspense page for react
